refactor: merge snippets/types.ts into autocomplete/types.ts
Consolidated autocomplete snippet type definitions into a single file:
- Moved AutocompleteSnippetType enum and all snippet interfaces to types.ts
- Updated 17 import statements across the codebase
- Deleted redundant snippets/types.ts file
