+++ src/services/continuedev/core/autocomplete/templating/filtering.ts
@@ -5,7 +5,7 @@ import {
 	AutocompleteSnippet,
 	AutocompleteSnippetType,
 	AutocompleteStaticSnippet,
-} from "../snippets/types"
+} from "../types"
 import { HelperVars } from "../util/HelperVars"
 import { formatOpenedFilesContext } from "./formatOpenedFilesContext"
