+++ src/services/ghost/classic-auto-complete/__tests__/GhostContextProvider.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, vi } from "vitest"
 import { getProcessedSnippets } from "../getProcessedSnippets"
 import { AutocompleteInput, GhostContextProvider } from "../../types"
-import { AutocompleteSnippetType } from "../../../continuedev/core/autocomplete/snippets/types"
+import { AutocompleteSnippetType } from "../../../continuedev/core/autocomplete/types"
 import { GhostModel } from "../../GhostModel"
 import { RooIgnoreController } from "../../../../core/ignore/RooIgnoreController"
 import crypto from "crypto"
