+++ src/test-llm-autocompletion/mock-context-provider.ts
@@ -12,7 +12,7 @@ import {
 	AutocompleteSnippetType,
 	type AutocompleteCodeSnippet,
 	type AutocompleteStaticSnippet,
-} from "../services/continuedev/core/autocomplete/snippets/types.js"
+} from "../services/continuedev/core/autocomplete/types.js"
 import type { RecentlyEditedRange } from "../services/continuedev/core/autocomplete/util/types.js"
 import type { ContextFile } from "./test-cases.js"
